--- MANIFEST.in
@@ -1,17 +1,0 @@
-include *.txt
-include *.md
-include *.cfg
-include *.pkl
-recursive-include ./datacube_bigmart/*
-
-
-include datacube_bigmart/datasets/train.csv
-include datacube_bigmart/datasets/test.csv
-include datacube_bigmart/trained_models/*.pkl
-
-
-include ./requirements.txt
-exclude *.log
-
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
