--- ml_api/api/__init__.py
@@ -1,3 +1,0 @@
-from ml_api.api.config import PACKAGE_ROOT
-
-
