--- ml_api/api/config.py
@@ -1,64 +1,0 @@
-import logging
-from logging.handlers import TimedRotatingFileHandler
-import os
-import sys
-from os.path import dirname, abspath, join
-
-PACKAGE_ROOT = abspath(dirname(__file__))
-
-FORMATTER = logging.Formatter(
-    "%(asctime)s — %(name)s — %(levelname)s —"
-    "%(funcName)s:%(lineno)d — %(message)s")
-LOG_DIR = join(PACKAGE_ROOT, 'logs')
-os.makedirs(LOG_DIR, exist_ok=True)
-LOG_FILE = join(LOG_DIR, 'ml_api.log')
-
-
-def get_console_handler():
-    console_handler = logging.StreamHandler(sys.stdout)
-    console_handler.setFormatter(FORMATTER)
-    return console_handler
-
-
-def get_file_handler():
-    file_handler = TimedRotatingFileHandler(
-        LOG_FILE, when='midnight')
-    file_handler.setFormatter(FORMATTER)
-    file_handler.setLevel(logging.WARNING)
-    return file_handler
-
-
-def get_logger(*, logger_name):
-    """Get logger with prepared handlers."""
-
-    logger = logging.getLogger(logger_name)
-
-    logger.setLevel(logging.INFO)
-
-    logger.addHandler(get_console_handler())
-    logger.addHandler(get_file_handler())
-    logger.propagate = False
-
-    return logger
-
-
-class Config:
-    DEBUG = False
-    TESTING = False
-    CSRF_ENABLED = True
-    SECRET_KEY = 'this-really-needs-to-be-changed'
-    SERVER_PORT = 5000
-
-
-class ProductionConfig(Config):
-    DEBUG = False
-    SERVER_PORT = os.environ.get('PORT', 5000)
-
-
-class DevelopmentConfig(Config):
-    DEVELOPMENT = True
-    DEBUG = True
-
-
-class TestingConfig(Config):
-    TESTING = True
