--- ml_api/tests/conftest.py
@@ -1,17 +1,0 @@
-import pytest
-
-from ml_api.api.app import create_app
-from ml_api.api.config import TestingConfig
-
-#Fixtures provide an easy way to setup and teardown resources
-@pytest.fixture
-def app():
-    app = create_app(config_object=TestingConfig)
-    with app.app_context():
-        yield app
-
-
-@pytest.fixture
-def flask_test_client(app):
-    with app.test_client() as test_client:
-        yield test_client
