+++ packages/ml_api/requirements.txt
@@ -1,0 +1,7 @@
+# api
+flask==1.0.2
+
+# local regression_model package
+# update with your local path
+-e "~/GitHub/bigmart/packages/datacube_bigmart/"
+
