--- requirements.txt
@@ -1,45 +1,0 @@
-appdirs==1.4.4
-attrs==20.2.0
-certifi==2020.6.20
-chardet==3.0.4
-click==7.1.2
-colorama==0.4.1
-distlib==0.3.1
-filelock==3.0.12
-Flask==1.1.2
-idna==2.10
-importlib-metadata==1.7.0
-iniconfig==1.0.1
-itsdangerous==1.1.0
-Jinja2==2.11.2
-joblib==0.17.0
-kaggle==1.5.8
-MarkupSafe==1.1.1
-numpy==1.19.2
-packaging==20.4
-pandas==1.1.3
-pluggy==0.13.1
-py==1.9.0
-pyasn1==0.4.8
-pyparsing==2.4.7
-pytest==6.1.1
-python-dateutil==2.8.1
-python-slugify==4.0.0
-pytz==2020.1
-requests==2.24.0
-rsa==3.4.2
-scikit-learn==0.23.2
-scipy==1.5.2
-simplejson==3.17.2
-six==1.15.0
-slugify==0.0.1
-text-unidecode==1.3
-threadpoolctl==2.1.0
-toml==0.10.1
-tox==3.20.0
-tqdm==4.43.0
-urllib3==1.24.3
-virtualenv==20.0.33
-Werkzeug==1.0.1
-xgboost==1.2.0
-zipp==3.3.0
