--- tox.ini
@@ -1,30 +1,0 @@
-[tox]
-envlist = datacube_bigmart
-skipsdist = True
-
-
-[testenv]
-install_command = pip install {opts} {packages}
-deps =
- 	-rrequirements.txt
-
-setenv =
-    PYTHONPATH=.
-
-commands =
-     python datacube_bigmart/train_pipeline.py
-#pytest tests/
-
-[testenv:install_locally]
-deps =
-    {[testenv]deps}
-
-setenv =
-    PYTHONPATH=.
-
-commands =
-    pip uninstall -y datacube_bigmart
-    #python setup.py sdist bdist_wheel
-    python setup.py install --force
-    python datacube_bigmart/train_pipeline.py
-
